refactor: use a class for common build code
